Use fixed-width types for baud rate and file offsets
baud_rate was an int, which cannot hold 115200 or 230400 on targets
where int is 16-bit (AVR/Arduino). Widen it to uint32_t and print the
uint32_t-typed fields with the inttypes.h PRIu32 macro, since uint32_t
is unsigned long on those toolchains and %d reads the wrong vararg
width.

+++ src/osdp_pd.c
@@ -1291,8 +1291,8 @@ static void osdp_pd_update(struct osdp_pd *pd)
 			comset_done_cmd.comset.baud_rate = pd->comset_pending.baud_rate;
 			do_command_callback(pd, &comset_done_cmd);
 			pd->address = (int)pd->comset_pending.address;
-			pd->baud_rate = (int)pd->comset_pending.baud_rate;
-			LOG_INF("COMSET Succeeded! New PD-Addr: %d; Baud: %d",
+			pd->baud_rate = pd->comset_pending.baud_rate;
+			LOG_INF("COMSET Succeeded! New PD-Addr: %d; Baud: %" PRIu32,
 				pd->address, pd->baud_rate);
 		}
 		osdp_phy_progress_sequence(pd);
